agent/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
trafilatura>=1.12.0
pydantic>=2.0.0
lancedb>=0.10.0
pyarrow
numpy
pandas